requirements.txt
telethon>=1.34.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
sqlalchemy>=2.0.0
pydantic>=2.5.0
simpleaudio>=1.0.4
python-dotenv>=1.0.0
loguru>=0.7.0
requests>=2.31.0
